implement crud for user

+++ src/controllers/UserController.ts
@@ -37,17 +37,7 @@ export class UserController {
         }
     }
 
-    async update(req: UpdateUserRequest, res: Response, next: NextFunction) {
-        // In our project: We are not allowing user to change the email id since it is used as username
-        // In our project: We are not allowing admin user to change others password
-
-        // Validation
-        const result = validationResult(req)
-        if (!result.isEmpty()) {
-            return res.status(400).json({ errors: result.array() })
-        }
-
-        const { firstName, lastName, role, email, tenantId } = req.body
+    async getById(req: Request, res: Response, next: NextFunction) {
         const userId = req.params.id as string
 
         if (isNaN(Number(userId))) {
@@ -55,9 +45,54 @@ export class UserController {
             return
         }
 
-        this.logger.debug('Request for updating a user', req.body)
+        try {
+            const user = await this.userService.findById(Number(userId))
+            if (!user) {
+                next(createHttpError(404, 'User not found.'))
+                return
+            }
+            // eslint-disable-next-line @typescript-eslint/no-unused-vars
+            const { password: _, ...userWithoutPassword } = user
+            res.json(userWithoutPassword)
+        } catch (err) {
+            next(err)
+        }
+    }
+
+    async getAll(req: Request, res: Response, next: NextFunction) {
+        try {
+            const queryParams = req.query
+            const [users, total] = await this.userService.getAll({
+                q: (queryParams.q as string) || '',
+                role: (queryParams.role as string) || '',
+                perPage: queryParams.perPage ? Number(queryParams.perPage) : 10,
+                currentPage: queryParams.currentPage
+                    ? Number(queryParams.currentPage)
+                    : 1,
+            })
+            res.json({ users, total })
+        } catch (err) {
+            next(err)
+        }
+    }
+
+    async update(req: UpdateUserRequest, res: Response, next: NextFunction) {
+        const userId = req.params.id as string
+
+        if (isNaN(Number(userId))) {
+            next(createHttpError(400, 'Invalid url param.'))
+            return
+        }
+
+        const { firstName, lastName, role, email, tenantId } = req.body
 
         try {
+            const existingUser = await this.userService.findById(Number(userId))
+            if (!existingUser) {
+                next(createHttpError(404, 'User not found.'))
+                return
+            }
+
             await this.userService.update(Number(userId), {
                 firstName,
                 lastName,
@@ -67,7 +102,6 @@ export class UserController {
             })
 
             this.logger.info('User has been updated', { id: userId })
-
             res.json({ id: Number(userId) })
         } catch (err) {
             next(err)
@@ -83,6 +117,12 @@ export class UserController {
         }
 
         try {
+            const existingUser = await this.userService.findById(Number(userId))
+            if (!existingUser) {
+                next(createHttpError(404, 'User not found.'))
+                return
+            }
+
             await this.userService.deleteById(Number(userId))
 
             this.logger.info('User has been deleted', {
